refactor(stats): small changes to pack rendering
