+++ src/hooks.server.ts
@@ -1,6 +1,6 @@
-import { init } from "$lib/custom_resources";
 import { updateCollections } from "$lib/server/constants/update-collections";
 import { updateItems } from "$lib/server/constants/update-items";
+import { init } from "$lib/server/custom_resources";
 import { parseNEURepository } from "$lib/server/helper/NotEnoughUpdates/parseNEURepository";
 import { updateNotEnoughUpdatesRepository } from "$lib/server/helper/NotEnoughUpdates/updateNEURepository";
 import { getPrices } from "skyhelper-networth";
