+++ static/resourcepacks/Vanilla/assets/minecraft/mcpatcher/cit/textures/leather_boots.properties
@@ -1,0 +1,4 @@
+type=item
+texture=leather_boots
+vanillaId=301
+damage=0
