+++ static/resourcepacks/Vanilla/assets/minecraft/mcpatcher/cit/textures/leather_chestplate.properties
@@ -1,0 +1,4 @@
+type=item
+texture=leather_chestplate
+vanillaId=299
+damage=0
