+++ static/resourcepacks/Vanilla/assets/minecraft/mcpatcher/cit/textures/leather_helmet.properties
@@ -1,0 +1,4 @@
+type=item
+texture=leather_helmet
+vanillaId=298
+damage=0
