+++ static/resourcepacks/Vanilla/assets/minecraft/mcpatcher/cit/textures/leather_leggings.properties
@@ -1,0 +1,4 @@
+type=item
+texture=leather_leggings
+vanillaId=300
+damage=0
